requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.47.0
google-cloud-storage
cloud-sql-python-connector==1.4.3
psycopg2-binary==2.9.9
pgvector==0.2.4
pypdf==3.17.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
streamlit==1.29.0
sqlalchemy==2.0.23
gunicorn==21.2.0
langchain
langchain-google-vertexai==0.1.3
numpy==1.24.3
pandas==2.1.4
